Restructuring and adding BME280

+++ src/globals.h
@@ -18,11 +18,13 @@ struct {
     float one;
     float two;
     float three;
-    int four;
+    float four;
+
     int five;
     int six;
     int seven;
     int eight;
+
     int nine;
     int ten;
     int rssi;
@@ -64,7 +66,7 @@ const char* version = xstr(VERSION);
 const char* clientId = xstr(CLIENTID);
 const char* mqttserver = xstr(MQTTSERVER);
 
-const char payloadFormat[] = "{\"s\":\"%s\", \"v1\":%f, \"v2\":%f, \"v3\":%f, \"v4\":%u, \"v5\":%u, \"v6\":%u, \"v7\":%u, \"v8\":%u, \"v9\":%u, \"v10\":%u, \"rssi\":%d}";
+const char payloadFormat[] = "{\"s\":\"%s\", \"v1\":%f, \"v2\":%f, \"v3\":%f, \"v4\":%f, \"v5\":%u, \"v6\":%u, \"v7\":%u, \"v8\":%u, \"v9\":%u, \"v10\":%u, \"rssi\":%d}";
 const char payloadFormat1[] = "{\"s\":\"%s\", \"v1\":%f, \"v2\":%f, \"v3\":%f, \"v4\":%u, \"v5\":%u, \"v6\":%u, \"v7\":%u, \"v8\":%u, \"v9\":%u, \"v10\":%u}";
 const char sdFormat[] = "%s,%f,%f,%f,%u,%u,%u,%u,%u,%u,%u,%d";
